fix: restore WASM compatibility via conditional imports
- shareText(): native delegates to share_plus, web uses the browser
  Web Share API (package:web) with the existing clipboard fallback
- NetworkNotifier: split into io implementation and web no-op stub;
  web never showed notifications (no web init settings), so behaviour
  is unchanged while dart:io leaves the web import graph
