+++ lib/src/notifications/network_notifier_web.dart
@@ -1,0 +1,31 @@
+import 'package:flutter/foundation.dart';
+
+import '../models/network_entry.dart';
+import 'alert_throttler.dart';
+
+/// Web stub of [NetworkNotifier]: a permanent no-op.
+///
+/// The notifier has never configured web initialization settings, so web
+/// builds have never shown notifications. This stub preserves that behaviour
+/// while keeping `flutter_local_notifications` (which transitively imports
+/// `dart:io`) out of the web import graph for WASM compatibility.
+class NetworkNotifier {
+  /// Creates a no-op notifier. Parameters mirror the native implementation so
+  /// call sites compile unchanged; [plugin] is accepted but never used.
+  NetworkNotifier({Object? plugin, AlertThrottler? throttler, this.onTap});
+
+  /// Invoked when the user taps the notification. Never fires on web.
+  final VoidCallback? onTap;
+
+  /// Always `false`: notifications are not supported on the web build.
+  bool get isAvailable => false;
+
+  /// No-op; the notifier stays unavailable.
+  Future<void> init() async {}
+
+  /// No-op.
+  Future<void> showOrUpdate(NetworkEntry entry, int totalCount) async {}
+
+  /// No-op.
+  Future<void> cancel() async {}
+}
